Moved plugin types to @grafana/ui
