+++ public/app/features/datasources/state/navModel.ts
@@ -1,4 +1,5 @@
-import { DataSource, NavModel, NavModelItem, PluginMeta } from 'app/types';
+import { DataSource, NavModel, NavModelItem } from 'app/types';
+import { PluginMeta } from '@grafana/ui/src/types';
 import config from 'app/core/config';
 
 export function buildNavModel(dataSource: DataSource, pluginMeta: PluginMeta): NavModelItem {
