+++ public/app/plugins/datasource/loki/datasource.ts
@@ -2,7 +2,8 @@ import _ from 'lodash';
 
 import * as dateMath from 'app/core/utils/datemath';
 import { LogsStream, LogsModel, makeSeriesForLogs } from 'app/core/logs_model';
-import { PluginMeta, DataQuery } from 'app/types';
+import { DataQuery } from 'app/types';
+import { PluginMeta } from '@grafana/ui/src/types';
 import { addLabelToSelector } from 'app/plugins/datasource/prometheus/add_label_to_query';
 
 import LanguageProvider from './language_provider';
