+++ public/app/plugins/datasource/testdata/QueryEditor.tsx
@@ -1,0 +1,12 @@
+import React, { PureComponent } from 'react';
+
+interface Props {
+}
+
+export class QueryEditor extends PureComponent<Props> {
+  render() {
+    return (
+      <h2>Test data</h2>
+    );
+  }
+}
